uv-unix: avoid exec-after-fork race in uv_spawn()
The race condition in a nutshell:

    if ((pid = fork()) > 0) {
      kill(pid, SIGTERM);
    }
    else if (pid == 0) {
      execve("/bin/cat", argp, envp);
    }

The parent sends a signal immediately after forking.
Since the child may not have called `execve()` yet,
there is no telling what process receives the signal,
our fork or /bin/cat.

To avoid ambiguity, we create a pipe with both ends
marked close-on-exec. Then, after the call to `fork()`,
the parent polls the read end until it sees POLLHUP.