build: serial-tests was added in automake v1.12
Commit a97685e added a check that tries to ascertain whether the
serial-tests option is supported by automake. It assumed that said
option was added in automake v1.11 but it turns out that's wrong, it
wasn't added until v1.12. Update the check.

Of course, none of this would have been necessary if the automake people
had simply added a macro that tells you if serial-tests is supported or
not.